+++ app/src/main/java/com/example/riseep3/ui/componenten/overview/ResultOutlinedField.kt
@@ -10,7 +10,7 @@ import androidx.compose.ui.Modifier
 
 @Composable
 fun ResultOutlinedField(
-    result: Int
+    result: Double
 ) {
     OutlinedTextField(
         value = "€$result",
